Add files via upload
ShipGrrr game opening scene with a 'Welcome to ShipGrrr!' message in the middle of the canvas depicting an ocean. There are three refuel slots at the top, the first one is bright green to denote it is an active slot to steer the ship to. The other two are darkened grey as they are not available for refueling. Three icebergs are randomly situated in the top ocean refuel zone. There are orcas and cachalots swimming in two different directions in the lanes. The Kraken's tentacles popped up randomly from underneath the ocean. The ship has a sun on its flag and is at the bottom starter safety zone ready to be navigated by the player.